Update the favicon for the search engine
It seems that the original favicon referenced by the favicon_url property no longer exists, causing firefox to show only a magnifier icon in the search bar. Thus, the property is changed to reference the favicon.ico file in the base directory of the website. Many websites use this file as a fallback favicon for older browsers, so it is not likely to be removed in the near future.